Create color choosing game
It is a fun game made from python language. It is made from conditional statements and looping statements.

To win in this game we must win against the computer. The color chosen by the player must match the color of the computer to earn a point.

The computer chooses a color using the random module. Winning the game against the computer is pragmatically difficult.